convert a string to an integer

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -8,7 +8,7 @@
 int _atoi(char *s)
 {
 	int sign = 1;
-	long int result = 0;
+	int result = 0;
 	int i = 0;
 	int len = 0;
 	int e = 0;
